research/requirements.txt
pydantic==2.5.2
pymongo==4.6.1
psycopg2-binary==2.9.9
pydantic-settings==2.1.0
Faker==20.1.0